requirements.txt
bilibili_api_python==15.4.4
openai==0.27.2
playsound==1.2.2
pygtrans==1.5.2
PyYAML==6.0
requests==2.31.0
websocket_client==1.5.3
